move giveth related files to giveth directory
